Work in progress to replace json-lib with Jackson
Jackson is the most performant Json parser, we have to use it. I also took a few minutes to cleanup bad commons-logging dependencies, it is so obsolete... instead, we have to rely on slf4j for the API, and logback for the logger implementation.

Also add a useful .gitignore file ;-)