Worker node discovery through JGroups
Worker node now announces themselves through a JGroups channel so that we can more easily detect them and detect their failure.

Unfortunately, I did not find any way to retrieve the distant node IP address from the received notification so I have to send it through the state transfer. This means that on a multi-homed machine, it might send the wrong IP address. Fortunately, this is not the case for this challenge :-)