Switch JGroups to full TCP for messages passing
Because multicast is not available we have to use a full TCP stack for messages. Meaning that messages to all members uses unicast now, unfortunately.

Note also that node discovery seems to not be very reliable, but at least it seems to work pretty well. If we want to stick with jgroups, we'll have to do some adaptations to use it in a better way.